Move ITreeStateContext
