+++ packages/lodestar/src/chain/chain.ts
@@ -26,7 +26,7 @@ import {TreeBacked} from "@chainsafe/ssz";
 import {AbortController} from "abort-controller";
 import {FAR_FUTURE_EPOCH} from "../constants";
 import {IBeaconDb} from "../db";
-import {ITreeStateContext, StateContextCache} from "./stateContextCache";
+import {StateContextCache} from "./stateContextCache";
 import {CheckpointStateCache} from "./stateContextCheckpointsCache";
 import {IBeaconMetrics} from "../metrics";
 import {notNullish} from "../util/notNullish";
@@ -35,7 +35,7 @@ import {BlockPool, BlockProcessor} from "./blocks";
 import {IBeaconClock, LocalClock} from "./clock";
 import {ChainEventEmitter} from "./emitter";
 import {handleChainEvents} from "./eventHandlers";
-import {IBeaconChain} from "./interface";
+import {IBeaconChain, ITreeStateContext} from "./interface";
 import {IChainOptions} from "./options";
 import {IStateRegenerator, QueuedStateRegenerator} from "./regen";
 import {LodestarForkChoice} from "./forkChoice";
