+++ packages/lodestar/src/chain/emitter.ts
@@ -3,8 +3,7 @@ import StrictEventEmitter from "strict-event-emitter-types";
 
 import {Attestation, Checkpoint, Epoch, SignedBeaconBlock, Slot, Version} from "@chainsafe/lodestar-types";
 import {IBlockSummary} from "@chainsafe/lodestar-fork-choice";
-import {ITreeStateContext} from "./stateContextCache";
-import {IBlockJob} from "./interface";
+import {IBlockJob, ITreeStateContext} from "./interface";
 import {AttestationError, BlockError} from "./errors";
 
 /**
